+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     * Bump `paypal-messages` to version `1.0.3`
 * ShopperInsights (BETA)
     * Add `PageType.PAYMENT_METHOD_SELECTION` and `PageType.PAYMENT_METHOD_ADDITION`
+    * Add `generateCustomerRecommendations` method to `ShopperInsightsClientV2`
 
 ## 5.10.0 (2025-04-29)